refactor: 2단계 1차 피드백 반영 - 번호 중복 여부 검증, 로또 검증 로직 Lotto로 이동
