+++ README.md
@@ -33,6 +33,7 @@
 - [x] 구입금액 입력
 - [x] 지난주 당첨번호 입력
   - [x] 번호 갯수 6개가 아닌 경우 IllegalArgumentException throw
+  - [x] 중복번호가 있는 경우 IllegalArgumentException throw
 - [x] 구매한 개수 출력
 - [x] 구매한 로또 목록 출력
 - [x] 당첨 통계 출력
